Modifiche all'util per il file system e all plot

+++ Hierarchical_directory/Util.py
@@ -10,6 +10,7 @@ from dataBase import *
 from curses import *
 from dataBase import dataBase
 from dataBase import dataBase
+from pathlib import Path
 import toPlotNetwork
 
 # Variabili globali
@@ -38,15 +39,15 @@ def ip_formatting(ipv4,ipv6,port):
     return p2p+pp2p
 
 def ip_deformatting(ip,port,ttl = None):
-	
+
     ipv4, ipv6 = ip.split('|')
-    
+
     f_ipv4 = re.sub('[.]0+','.',ipv4)
-    
+
     f_ipv6 = str(ipa.ip_address(ipv6))
 
     f_port = int(port)
-    
+
     if(ttl != None):
         f_ttl = int(ttl)
         return f_ipv4, f_ipv6, f_port, f_ttl
@@ -70,8 +71,10 @@ def initializeFiles():
     f.write('#### Log file launched on {:%d-%m-%Y %H:%M:%S} ####'.format(datetime.now())+'\n\n')
     f.close()
 
-    file = open('File_System.txt', "w")
-    file.close()
+    file = Path("File_System.txt")
+    if(not file.is_file()):
+        file = open('File_System.txt', "w")
+        file.close()
 
 def printLog(desc):
 
@@ -86,7 +89,7 @@ def printError(desc):
     f.close()
 
 def menu(stdscr, listMenu, titleMenu, flag = None):
-    
+
     attr = {}
     numset = []
 
@@ -104,7 +107,7 @@ def menu(stdscr, listMenu, titleMenu, flag = None):
     init_pair(3, COLOR_RED, COLOR_BLACK)
 
     use_default_colors() # Prendi i colori del terminale
-    
+
     attr['normal'] = color_pair(1)
     attr['highlighted'] = color_pair(2)
     attr['system'] = color_pair(3)
@@ -115,7 +118,7 @@ def menu(stdscr, listMenu, titleMenu, flag = None):
 
     depth = 0 # Profondità del menu per selezionare il titolo
     c=0
-    option = 0 
+    option = 0
     while c != 10: # enter --> esci
 
         stdscr.clear()
@@ -132,7 +135,7 @@ def menu(stdscr, listMenu, titleMenu, flag = None):
                     if i in [len(listMenu) - 4, len(listMenu) - 2]:
                         par = attr['system']
                     else:
-                        par = attr['normal']        
+                        par = attr['normal']
                 else:
                     if i == len(listMenu) - 2:
                         par = attr['system']
@@ -147,12 +150,12 @@ def menu(stdscr, listMenu, titleMenu, flag = None):
                 else:
                     par = attr['normal']
 
-            stdscr.addstr('\n\t[{0}] '.format(count))         
+            stdscr.addstr('\n\t[{0}] '.format(count))
             stdscr.addstr(listMenu[i] + '\n', par)
             count += 1
-        
+
         c = stdscr.getch()
-        
+
         if c == KEY_UP:
 
             if option > 0:
@@ -171,11 +174,11 @@ def menu(stdscr, listMenu, titleMenu, flag = None):
             option = (c - 49) * 2
 
         elif c == 10:
-            
+
             if type(listMenu[option + 1]) != list: # L'elemento non è una lista quindi è quello che cerco
-               
+
                 return listMenu[option + 1]
-            
+
             else:                                  # L'elemento è una lista, quindi prevedo un altro livello di menu
 
                 lenMenu = len(listMenu[option + 1])
@@ -184,7 +187,7 @@ def menu(stdscr, listMenu, titleMenu, flag = None):
                     depth -= 1
                     depthMenu.remove(listMenu)
 
-                    swpList = listMenu[option + 1] 
+                    swpList = listMenu[option + 1]
 
                     listMenu.remove(listMenu[option + 1])
                     listMenu.remove('Back')
@@ -194,13 +197,13 @@ def menu(stdscr, listMenu, titleMenu, flag = None):
                         listMenu.remove(None)
 
                     listMenu = swpList # Swap dei due menu
-                
+
                 else: # Menu inferiore
                     depth += 1
 
                     listMenu[option + 1].append('Back')
                     listMenu[option + 1].append(listMenu)
-                    
+
                     if flag != None:
                         listMenu[option + 1].append('Abort')
                         listMenu[option + 1].append(None)
@@ -223,26 +226,35 @@ def statusNetwork():
     nodes = [] #peer della rete
     sp_list = [] #superpeers della rete
     list_sp = db.retrieveSuperPeers()
+    Util.printLog("LIST_SP " + str(list_sp))
     for lsp in list_sp:
-        lsp_d = ip_deformatting(lsp[0], lsp[1]) 
+        lsp_d = ip_deformatting(lsp[0], lsp[1])
         nodes.append(lsp_d[0])
         sp_list.append(lsp_d[0])
     my_ip = db.retrieveConfig(('selfV4', 'selfV6'))
-    ip_SP = my_ip.selfV4 + "|" + my_ip.selfV6
+    ip_SP = my_ip.selfV4
 
     logged_list = [] # peer logged to supernode
     logged = dbs.retrieveLOGINsp()
+    Util.printLog("LOGGED_LIST " + str(logged_list))
     for lg in logged:
         lg_d = ip_deformatting(lg[0],lg[1])
         nodes.append(lg_d[0])
         logged_list.append(lg_d[0])
 
-    nodes.insert(0,ip_SP[:15])
+    seen = set()
+    uniq = []
+    for x in nodes:
+        if x not in seen:
+            uniq.append(x)
+            seen.add(x)
+
+    nodes.insert(0,ip_SP)
 
     edges = [] # archi della rete
     sol = [] # archi soluzione (traffico)
     for e1 in sp_list:
-        edge = (ip_SP[:15],e1)
+        edge = (ip_SP,e1)
         edges.append(edge)
     for e2 in logged_list:
         edge = (e2, ip_SP[:15])
@@ -250,11 +262,12 @@ def statusNetwork():
 
     num_sp = len(list_sp)
     num_peer = len(sol)
-    toPlotNetwork.toPlot(nodes, edges, sol, num_sp, num_peer)
+    Util.printLog("ROBA DEL PLOT: " + "---" + str(seen) + "---" + str(edges))
+    toPlotNetwork.toPlot(seen, edges, sol, num_sp, num_peer)
 
 
 def updatePeers():
-    
+
     globalLock.acquire()
     mode = Util.mode
     Util.mode = 'update'
@@ -263,12 +276,12 @@ def updatePeers():
     lock.acquire()
     db = dataBase()
 
-    listNormal = db.retrievePeers() 
+    listNormal = db.retrievePeers()
     listSuper = db.retrieveSuperPeers()
 
     #db.deletePeers()
     db.deleteSuperPeers()
-    
+
     config = db.retrieveConfig(('selfV4','selfV6','selfP','ttl'))
     idPacket = ip_packet16()
     ip = ip_formatting(config.selfV4, config.selfV6, config.selfP)
@@ -291,14 +304,14 @@ def updatePeers():
 
         ipv4, ipv6, port = Util.ip_deformatting(peer[0],peer[1])
         con = Conn(ipv4, ipv6, port)
-        
+
         if con.connection():
 
             count += 1
             con.s.send(pack.encode())
             printLog("Richiesta SUPE a vicino ::: " + ipv4)
             con.deconnection()
-        
+
         else:
             printLog("Richiesta SUPE fallita per ::: " + ipv4)
